Remove packet-print.h/cpp which duplicates key info printing functionality
