+++ include/rekey/rnp_key_store.h
@@ -154,8 +154,6 @@ size_t     rnp_key_store_get_key_count(const rnp_key_store_t *);
 pgp_key_t *rnp_key_store_get_key(const rnp_key_store_t *, size_t);
 list       rnp_key_store_get_keys(const rnp_key_store_t *);
 
-bool rnp_key_store_list(FILE *fp, const rnp_key_store_t *, const int);
-
 pgp_key_t *rnp_key_store_add_key(rnp_key_store_t *, pgp_key_t *);
 
 bool rnp_key_store_remove_key(rnp_key_store_t *, const pgp_key_t *);
