+++ src/librepgp/repgp.cpp
@@ -42,7 +42,6 @@
 #include "stream-key.h"
 
 #include "internal_types.h"
-#include "packet-print.h"
 #include "utils.h"
 #include "crypto.h"
 #include "pgp-key.h"
